Remove unused variables/imports, remove cross pages components imports
