+++ src/app/settings/(general)/_components/GeneralSettings.tsx
@@ -1,7 +1,6 @@
 "use client";
 
 import React from "react";
-import styled from "styled-components";
 
 import { useTRPC } from "@/trpc/client";
 import { useQuery } from "@tanstack/react-query";
@@ -9,13 +8,6 @@ import { useQuery } from "@tanstack/react-query";
 import UsageSettings from "./UsageSettings";
 import LLMSettings from "./LLMSettings";
 import PreferencesSettings from "./PreferencesSettings";
-
-export const BorderWrapper = styled.div`
-  border: var(--measurement-small-30) solid var(--font-color-alpha-10);
-  border-radius: var(--measurement-medium-30);
-
-  background: var(--contrast-color);
-`;
 
 function GeneralSettings() {
   const trpc = useTRPC();
