+++ src/components/border-wrapper/index.ts
@@ -1,0 +1,10 @@
+"use client";
+
+import styled from "styled-components";
+
+export const BorderWrapper = styled.div`
+  border: var(--measurement-small-30) solid var(--font-color-alpha-10);
+  border-radius: var(--measurement-medium-30);
+
+  background: var(--contrast-color);
+`;
